requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0

# LLM Providers
openai>=1.77.0  # for DeepSeek API (OpenAI-compatible) and deepsearcher

# Observability
langfuse>=3.10.0

# RAG & Embeddings
chromadb==0.4.24
sentence-transformers==2.7.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0

# Utils for tokenization
tiktoken==0.7.0

# Data handling
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
PyYAML==6.0.1

# Integrations
google-api-python-client==2.127.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-telegram-bot==21.1.1

# Database
sqlalchemy==2.0.30
alembic==1.13.1

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Utils
aiohttp==3.9.5
httpx==0.27.0
python-dateutil==2.9.0
pytz>=2024.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0

# Development
black==24.4.2
flake8==7.0.0
mypy==1.10.0
